uodate lai

+++ DATN_SD22/src/main/java/com/poly/sneaker/repository/KhachHangRepository.java
@@ -1,13 +1,14 @@
 package com.poly.sneaker.repository;
 
 import com.poly.sneaker.entity.KhachHang;
-import com.poly.sneaker.entity.NhanVien;
+import org.hibernate.query.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.awt.print.Pageable;
 import java.util.List;
 
 @Repository
-public interface KhachHangRepository extends JpaRepository<KhachHang,Integer> {
+public interface KhachHangRepository extends JpaRepository<KhachHang,Long> {
     List<KhachHang> findByTen(String ten);
 }
